+++ app/assets/javascripts/kassi.js
@@ -74,21 +74,6 @@ function add_validator_methods() {
     );
 
   $.validator.
-    addMethod("required_when_not_neutral_feedback",
-      function(value, element, param) {
-        if (value == "") {
-          var radioButtonArray = new Array("1", "2", "4", "5");
-          for (var i = 0; i < radioButtonArray.length; i++) {
-            if ($('#grade-' + radioButtonArray[i]).is(':checked')) {
-              return false;
-            }
-          }
-        }
-        return true;
-       }
-    );
-
-  $.validator.
     addMethod( "positive_integer",
       function(value, element, param) {
         var n = ~~Number(value);
@@ -368,7 +353,6 @@ function initialize_accept_transaction_form(
   minimum_price_message) {
 
   auto_resize_text_areas("text_area");
-  style_action_selectors();
 
   if (commission_percentage != null) {
     if (form_type === "simple") {
@@ -472,49 +456,9 @@ function update_complex_form_price_fields(commissionPercentage, serviceFeeVat) {
   $("#total").text(youWillGet.toFixed(2) + euro);
 }
 
-function initialize_confirm_transaction_form() {
-  style_action_selectors();
-}
-
-function style_action_selectors() {
-  $(".conversation-action").each(function() {
-    $(this).find('label').hide();
-    $(this).find('.conversation-action').each(
-      function() {
-        $(this).removeClass('hidden');
-        $(this).click(
-          function() {
-            var action = $(this).attr('id');
-            $(this).siblings().removeClass('accept').removeClass('reject').removeClass('confirm').removeClass('cancel');
-
-            // Show or hide description text
-            $(".confirm-description").addClass('hidden');
-            $(".cancel-description").addClass('hidden');
-            $("." + action + "-description").removeClass('hidden');
-
-            // Show or hide price field
-            $(".conversation-price").addClass('hidden');
-            $("." + action +  "-price").removeClass('hidden');
-
-            // Show or hide payout details missing information
-            $(".hidden-accept-form").addClass('hidden');
-            $(".visible-when-" + action).removeClass('hidden');
-
-            $(this).addClass(action);
-            $(".conversation-action").find('input:radio[id=' + $(this).attr('name') + ']').attr('checked', true);
-            $("#conversation_message_attributes_action").val(action);
-            $("#conversation_status").val(action + 'ed');
-          }
-        );
-      }
-    );
-  });
-}
-
 function initialize_give_feedback_form(locale, grade_error_message, text_error_message) {
   auto_resize_text_areas("text_area");
   $('textarea').focus();
-  style_grade_selectors();
   var form_id = "#new_testimonial";
   $(form_id).validate({
     errorPlacement: function(error, element) {
@@ -534,24 +478,6 @@ function initialize_give_feedback_form(locale, grade_error_message, text_error_m
     submitHandler: function(form) {
       disable_and_submit(form_id, form, "false", locale);
     }
-  });
-}
-
-function style_grade_selectors() {
-  $(".feedback-grade").each(function() {
-    $(this).find('label').hide();
-    $(this).find('.grade').each(
-      function() {
-        $(this).removeClass('hidden');
-        $(this).click(
-          function() {
-            $(this).siblings().removeClass('negative').removeClass('positive');
-            $(this).addClass($(this).attr('id'));
-            $(".feedback-grade").find('input:radio[id=' + $(this).attr('name') + ']').attr('checked', true);
-          }
-        );
-      }
-    );
   });
 }